@@ -12,7 +12,6 @@
   "dependencies": {
     "@popperjs/core": "^2.0.6",
     "@stripe/stripe-js": "^1.3.0",
-    "@vue/compat": "^3.1.0",
     "core-js": "^3.6.4",
     "feather-icons": "^4.26.0",
     "frappe-charts": "^1.3.0",
@@ -21,7 +20,7 @@
     "luxon": "^1.22.0",
     "register-service-worker": "^1.6.2",
     "socket.io-client": "^2.4.0",
-    "vue": "^3.1.0",
+    "vue": "^3.2.16",
     "vue-router": "^4.0.5"
   },
   "devDependencies": {
