@@ -9,14 +9,16 @@ export default {
 		UpdateBillingDetails: () => import('@/components/UpdateBillingDetails.vue')
 	},
 	resources: {
-		prompts: {
-			method: 'press.api.account.user_prompts',
-			auto: true,
-			validate() {
-				if (document.cookie.includes('user_id=Guest')) {
-					return 'Not logged in';
+		prompts() {
+			return {
+				method: 'press.api.account.user_prompts',
+				auto: true,
+				validate() {
+					if (document.cookie.includes('user_id=Guest')) {
+						return 'Not logged in';
+					}
 				}
-			}
+			};
 		}
 	},
 	data() {
